requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.4
pytimeparse==1.1.8
openai==1.35.13